Fix code comments in generated cs
CppSharp has some _interesting_ conventions around comments. There
doesn't seem to be a way to have a plain, simple, raw comment (dude, I
can create a string better than it can generate one for me), so this
commit makes the comments in the cs file match what the comment should
be as far as the CppSharp.CsCommentPrinter seems to allow.